更新

+++ src/main/java/com/freelycar/saas/project/controller/ProjectController.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.Map;
 
 @Api(value = "门店项目管理", description = "门店项目管理接口", tags = "门店项目管理接口")
@@ -59,7 +60,7 @@ public class ProjectController {
     @ApiOperation(value = "修改项目列表位置", produces = "application/json")
     @PostMapping(value = "/switchLocation")
     @LoggerManage(description = "调用方法：修改项目列表位置")
-    public ResultJsonObject switchLocation(@RequestBody Map<String, Long> map) {
+    public ResultJsonObject switchLocation(@RequestBody Map<String, BigInteger> map) {
         boolean result = projectService.switchLocation(map);
         if (result) {
             return ResultJsonObject.getDefaultResult(null);
