+++ src/main/java/com/freelycar/saas/project/entity/Store.java
@@ -11,6 +11,7 @@ import org.hibernate.validator.constraints.Length;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Time;
 import java.sql.Timestamp;
 
@@ -70,7 +71,8 @@ public class Store implements Serializable {
     private String remark;
 
     @Column
-    private Long sort;
+    private BigInteger sort;
+
 
     @Column
     private String headUrl;
